@SevroAuBarca: Solucionando bugs

+++ frontend/src/app/Components/header/header.component.ts
@@ -1,6 +1,7 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { Router } from '@angular/router';
 import { LocalStorageService } from 'src/app/services/local-storage.service';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-header',
@@ -13,6 +14,8 @@ export class HeaderComponent implements OnInit {
     username: 'XD',
     id: '1',
   };
+  @ViewChild('options') options!: ElementRef<HTMLInputElement>;
+
   constructor(
     private _localStorage: LocalStorageService,
     private router: Router
@@ -27,8 +30,30 @@ export class HeaderComponent implements OnInit {
   }
 
   cerrarSesion() {
-    this._localStorage.deleteLocalStorage();
+    Swal.fire({
+      title: '¿Estas Seguro?',
+      text: 'Cerraras la sesion actual',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      cancelButtonText: 'Cancelas',
+      confirmButtonText: 'Cerrar Sesion',
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.toggleOptions();
+        this._localStorage.deleteLocalStorage();
 
-    this.router.navigate(['/']);
+        this.router.navigate(['/']);
+      }
+    });
+  }
+
+  toggleOptions() {
+    if (this.options.nativeElement.classList.contains('active')) {
+      this.options.nativeElement.classList.remove('active');
+    } else {
+      this.options.nativeElement.classList.add('active');
+    }
   }
 }
